feat: Emulate `TypeBound`s on parameters via constraints. (#1624)
This PR translates some `TypeBound`s in `hugr-core` to the `nonlinear`
constraint in `hugr-model`. This translation only occurs on parameters
that take a runtime type directly.

As a driveby change before the model stabilises, this PR also moves
constraints out of the parameter lists into their own list. In its
previous form this could have led to confusions about which parameter a
local variable index refers to when a constraint is situated between two
parameters in the list. We also remove constraints from aliases for now.

Closes https://github.com/CQCL/hugr/issues/1637.