fix: Resolve types in `Value`s and custom consts (#1779)
Collects and resolves extensions in the types stored inside a `Value`.
This includes the specified type of a `Sum` and other cached types.

Since we expect `CustomConst`'s `get_type` method to always return
signatures computed by a binary definition, I'd say we close #1742 as
not needed.

For some reason there's a random test on `hugr-py` that starts failing
when we enable this:
```
=================================== FAILURES ===================================
______________________________ test_higher_order _______________________________
Error parsing package: Error resolving opaque operation: Error in signature of operation 'prelude.Noop' in Node(3): Type arguments of node did not match params declared by definition: Wrong number of type arguments: 0 vs expected 1 declared type parameters
```

The fix for #1774 that I'm submitting immediately after this PR fixes it
so I'm skipping the test in this PR 🤷
I'll open an issue to investigate further after we make the release

---------

Co-authored-by: Seyon Sivarajah <seyon.sivarajah@quantinuum.com>