feat: Cleanup `Display` of types and arguments (#1802)
Updates the `Display` implementation of poly functions, functions,
types, type arguments, and type params so its easier to read.

Comparison below. I added the newlines manually.

Before:
```swift
forall Any _0.map(|i|i.to_string()).unwrap_or("-".to_string()).
    [array([
        Variable { v: TypeArgVariable { idx: 1, cached_decl: BoundedNat { bound: UpperBound(None) } } },
        Type { ty: TypeBase(Sum(General { rows: [TypeRowBase { types: [] }, TypeRowBase { types: [TypeBase(Variable(0, Any), Any)] }] }), Any) }
    ])]
 -> [[]][[
    array([
            Variable { v: TypeArgVariable { idx: 1, cached_decl: BoundedNat { bound: UpperBound(None) } } },
            Type { ty: TypeBase(Sum(General { rows: [TypeRowBase { types: [] }, TypeRowBase { types: [TypeBase(Variable(0, Any), Any)] }] }), Any) }
        ]),
        int([BoundedNat { n: 6 }])
    ]]
```
(yes, that is some hardcoded rust source)

After:
```swift
∀ (var#0 : Type) (var#1 : Nat). [array(var#1, Type([]+[var#0]))]
                             -> [[array(var#1, Type([]+[var#0])), int(6)]]
```

That is, a generic function on a type var and an unbounded natural that
takes an array of (nat, option type) and returns a tuple with the array
and an int.

It may be better if the tuple display used `()` instead of `[]`, but
that can be discussed later.

Update after review suggestions:
```swift
∀ (#0 : Type) (#1 : Nat). [array(#1, []+[#0])] -> [[array(#1, []+[#0]), int(6)]]
```