chore!: Update to `portgraph 0.13` / `petgraph 0.7` (#1878)
The new portgraph release comes with some perf improvements.
We require this update for #1869.

Closes #1872.

I added a benchmark variation for subgraphs with few nodes as a
drive-by.
This got improved by the new portgraph (skipping a full graph
traversal), but the runtime is still mostly the `O(n)` full-graph
traversal required for the convexity checking when building the
subgraph.

BREAKING CHANGE: Bumped public dependency `portgraph` to breaking
version `0.13`.