feat: Do not require `capnp` to be installed to compile `hugr-model` (#1907)
Now we do the same as with the schema and extensions definition; provide
a `just` command to update it manually and check on CI that things are
up-to-date.

This lets us remove the `build.rs` from `hugr-model`, removing the need
for users to have `capnproto` installed to compile our crates.

Closes #1667