chore: Add Cargo.lock, fix builds on 1.75 (#1942)
[icu4x](https://github.com/unicode-org/icu4x) bumped its MSRV to 1.81
for all of it's sub-crates in a minor release. This means that we can
[no longer build hugr with
1.75](https://github.com/CQCL/hugr/actions/runs/13569763210/job/37931524942#step:7:315).

This PR unignores the `Cargo.lock` and `cargo update`s the dependencies
to downgrade them to a supported version. We should see the 1.75 CI job
fail if dependabot tries to update them again.

This problem is caused by the dependency resolver v2 ignoring the
`rust-version` field in the crates. In rust 2024 the new v3 resolver
should be able to use only valid versions, but we have a long way until
we are able to use that.