feat: Rewrite for inlining a single Call (#1934)
Including adding a new method for copying descendants of another node
(preserving edges incoming to that subtree), with optional Substitution.

The choice of copying descendants rather than a subtree and its root may
seem a little strange. The minor difference is that in most usecases
(others are peeling tail loops, and monomorphization which can be
refactored to use this too) we're gonna update the root node optype too,
so we might as well just make the new optype first rather than
copy+overwrite. The bigger differences are
* Edges between the copied subtree and the root (e.g. recursive Calls)
want to point to the original FuncDefn, not the inlined copy (which will
be a DFG). This helps the inline-call usecase, and may help
monomorphization (recursive calls may have different type args, so we
should monomorphize their targets too), and makes no difference for loop
peeling (there are no edges to/from a TailLoop from inside).
* Edges already existing to the new root (the old Call node) can be left
untouched, rather than having to be moved (and the old Call node
deleted). Again I think this makes no difference for loop peeling, nor
monomorphization.

closes #1833, #1886

---------

Co-authored-by: Douglas Wilson <douglas.wilson@quantinuum.com>