ci: Ensure that lints checks run when hugr-llvm gets modified (#1950)
The `llvm` ci changes was independent of the general `rust` one, but CI
wouldn't run basic checks unless the former was active.

This caused the checks to be skipped on
https://github.com/CQCL/hugr/pull/1946, and the lint errors only
appeared after the merge.

Now the file changes flag hierarchy is
```mermaid
graph LR
    rc[rust-config] --> model
    rc --> llvm
    rc --> rust
    model --> rust
    llvm --> rust
    rust --> python
```
Most checks run on `rust` changes, but tests in the llvm crate are
skipped unless `llvm` is set.