feat: `hugr-model` AST (#1953)
This PR introduces the AST representation of `hugr-model`, which is less
efficient but more ergonomic than the tabled variant. It is used for
parsing and pretty printing, and can be converted from and into the
tabled representation. Individual parts, such as single terms, can be
parsed and printed which will help resolve #1600. The AST representation
can also become useful for Python bindings.

* Parsing used to be source -> table, but now source -> AST (-> table)
* Binary representation still loads directly into table form
* table `View`s itself as AST; tree form `resolve`s itself into table
* `struct Literal` factored out from `enum table::Term` and shared with
`enum syntax::Term`
* Model `Node`s no longer include params, these are now put inside an
`Apply` inside the node.
* Alias values therefore no longer in params but made part of
`Operation::DefineAlias`.
* `ListPart` and `TuplePart` combined into `SeqPart`