fix: Remove return from val_or_panic (#1999)
We shouldn't have been emitting a `ret` in `val_or_panic` - this leads
to behaviour where code that comes after the `val_or_panic` section
isn't emitted.

For example, the `test_s_to_u` test has been altered to add a constant
to the number after its converted. Before this fix it would always fail
because the addition was skipped.

---------

Co-authored-by: Douglas Wilson <141026920+doug-q@users.noreply.github.com>