fix: Inconsistent behaviour in `SiblingSubgraph::from_nodes` (#2011)
`SiblingSubgraph::from_nodes` and `from_node` have different
definitions.

The former takes a set of nodes and defines a subgraph where the inputs
and outputs correspond to all edges connecting the chosen nodes to
external ones.
When a node has a disconnected port, it is ignored.

The latter does not check edges. It takes a single node and creates a
subgraph using the dataflow signature of the operation.

Both definitions are valid and useful on their own, when calling
`from_nodes` with a single-element vector we redirected the call to
`from_node`, resulting in different port definitions.

This PR fixes that.