feat: ReplaceTypes pass allows replacing extension types and ops (#1989)
Also added `HugrMut::optype_mut` and `Type::as_extension`.

There is a slight naming issue in that this also allows replacing ops
(and consts), as is necessary to maintain validity of the Hugr.

I intend to deprecate the old `hugr_passes::lower` (also OpDef
`lower_funcs` and `enum LowerFunc`) in a follow-up PR.

This PR leaves things that If you replace a Copyable type with a Linear
one, you'll either get a SignatureError (if this results in a type like
`ArgMustBeCopyable[Qubit]`) or an invalid Hugr. (Or you might get lucky,
if every outgoing port of the type happens to have one successor!)
Adding `dup` / `free` follows in #2018.

- [X] TODO: test op-lowering Fn returning None