feat: Remove extension sets from `hugr-model`. (#2031)
This PR removes extension sets from `hugr-model` (see #1906).
`hugr-core` or `hugr-py` continue to have extension sets for now, but
those are ignored by export and imported as empty. This is important to
integrate before stabilising `hugr-model`, since it changes the arity of
various core constructors and operations (such as `core.fn` and
`core.call`).