chore: Avoid linking libpython for rust tests (#2035)
The `pyo3` feature in hugr-model is used for building the python package
extension for `hugr-py`. In general, this should disable linking against
`libpython` as indicated in the pyo3 docs:
-
https://pyo3.rs/main/building-and-distribution#the-extension-module-feature

Unfortunately, when building rust binaries that depend on pyo3 without
using maturin we _do_ need to link libpython as otherwise we end up with
missing symbol errors. [This includes builds for `cargo test` or `cargo
bench`](https://pyo3.rs/v0.24.0/faq#i-cant-run-cargo-test-or-i-cant-build-in-a-cargo-workspace-im-having-linker-issues-like-symbol-not-found-or-undefined-reference-to-_pyexc_systemerror).

Ok then, we could keep linking to libpython for those targets.
But alas, when using a managed python installation with `uv` sometimes
`libpython` does not get found by the linker. There's a miriad of
related issues, take your pick:
- libpython not available in uv's venv:
[astral-sh/uv#6812](https://www.github.com/astral-sh/uv/issues/6812)
- > This is not a bug - venvs are primarily *run*-time (as opposed to
development-time) environments

  https://bugs.python.org/issue43334#msg388720
- "A uv Environment Is Not Compatible With pyo3 Out Of The Box"
  [astral-sh/uv#11006](https://www.github.com/astral-sh/uv/issues/11006)
- A collection of related issues: [astral-sh/uv#11234
(comment)](https://www.github.com/astral-sh/uv/issues/11234#issuecomment-2635828351])
- etc.

There may be a way to fix this here, but for now as a temporary solution
I enabled `pyo3:extension-module` on `hugr-model` (so it can be built
without linking) and excluded that feature from the test commands in
`pre-commit` and `just`. With this the pre-commit checks run again on my
local dev env.

We should investigate this further later, and try removing the
`extension-module` option from `hugr-model`, enabling it only on maturin
builds.