feat!: Expand SimpleReplacement::ApplyResult type (#2085)
It is currently impossible to obtain the map from nodes in the
replacement graph to newly inserted nodes in `self` from the
`ApplyResult` type for `SimpleReplacement`.

This PR fixes this.

BREAKING CHANGE: The `ApplyResult` associated type of
`SimpleReplacement` is now a custom struct. The deleted nodes returned
previously can be obtained using the `result.deleted_nodes` field.