feat!: Update array LLVM lowering to use malloc (#2112)
Closes #2067

An `array<n, T>` is now lowered to a struct `{ptr, usize}` where `ptr`
is a heap allocated pointer of size at least `n * sizeof(T)` and the
`usize` is an offset pointing to the first element (i.e. the first
element is at `ptr + offset * sizeof(T)`).

The rational behind the additional offset is the `pop_left` operation
which bumps the offset instead of mutating the pointer. This way, we can
still free the original pointer when the array is discarded after a pop.

BREAKING CHANGE: Removed old lowering for value arrays