feat!: Bump MSRV to 1.85 (#2136)
Bumps the minimum supported rust version to 1.85, ~and migrates to the
2024 edition~.

Most of the changes here are automated. The diff is quite noisy due to
some changes in formatting, and me using this opportunity to auto--fix
some optional clippy lints. It may be easier to check the changes
per-commit.

EDIT: Edition change has been left for a separate PR, as it's quite
noisy.

BREAKING CHANGE: Bumped MSRV to 1.85