chore: Remove stray rewrite.rs file (#2142)
Oupsie, during one of the merge conflict resolutions I must have
forgotten to remove the old `rewrite.rs` file. As you can see in the
`hugr-core/src/hugr.rs` file, this is no longer a module and thus the
file should be deleted. It has been renamed to `patch.rs` in #2070