feat!: Accept outgoing ports in SimpleReplacement nu_out (#2151)
Currently, SimpleReplacement stores its output boundary map nu_out by
referring to nodes outside the deleted subgraph in the host graph. This
forces the invalidation set of the replacements to include nodes past
the output, forbidding simultaneous adjacent replacements.

This PR fixes this by allowing the keys of `nu_out` (i.e. the ports on
the output boundary of the subgraph) to be either incoming ports (as
before), or outgoing ports (in which case this is equivalent to
specifying the map on all incoming ports linked to the given outgoing
ports). The latter is less general but covers most use cases and reduces
the size of the invalidation set.

Closes https://github.com/CQCL/hugr/issues/2098

BREAKING CHANGE: Generalised arguments to [`SimpleReplacement::new`] and
[`SimpleReplacement::map_host_output`] to allow for outgoing ports.
Previous type signatures remain valid, however, type inference may fail.

---------

Co-authored-by: Mark Koch <48097969+mark-koch@users.noreply.github.com>