feat!: Return a node mapping in HugrInternals::region_portgraph (#2164)
Closes #2158

Annoyingly, we still need to return a `FlatRegion` (or some other
concrete wrapper) here so that _references_ to the graph implements some
of petgraph's traits (like `IntoNodeIdentifiers`) which are normally
implemented for `&'_ G`. There's no way to write that in a GAT
constraint :/
We'll have to wait for the big petgraph rewrite to clean that.

BREAKING CHANGE: `HugrInternals::region_portgraph` now also returns a
node mapping between hugr and portgraph nodes

---------

Co-authored-by: Mark Koch <48097969+mark-koch@users.noreply.github.com>