fix!: Use unique region_portgraph in convexity check (#2192)
Previously, `region_portgraph` was called twice: once to construct the
convexity checker (e.g. in `try_from_nodes`), and once to actually
construct the subgraph (in `try_new_with_checker`). As a result, if the
results from `region_portgraph` were not deterministic, the checker
would not be valid and the convexity check might fail. I've had to
narrow down the argument type, but there is currently only one type of
convexity checker used, so this should not impact anyone.

BREAKING CHANGE: `SiblingSubgraph` only accepts `TopoConvexChecker` as
convexity checker type