fix: Panic when an extension name is too long. (#2198)
This PR fixes a bug that led to a panic in import when the name of an
extension exceeded 23 bytes.

This was caused by a misreading of `IdentList::new_unchecked`: this
method not only avoids checking the identifier for validity but also
assumes that the input will fit into the 23 byte inline capacity of a
`SmolStr`.