feat!: Remove boundary map in SimpleReplacement (#2208)
`SimpleReplacement` is now so much simpler!!

I've just gone ahead and removed the `nu_inp` and `nu_out` maps of
`SimpleReplacement`. Makes defining replacements so much less verbose!
But you will probably (rightfully!) be sceptical that this can work.

The crux is: there is already a boundary map defined in the `subgraph:
SiblingSubgraph`. My claim is: **We can restrict `nu_inp` and `nu_out`
to be bijective maps between the boundary as defined by
`SiblingSubgraph` and the ports at the Input/Output nodes of the
replacement**. And once the map is bijective, we might as well restrict
it to the canonical identity map, mapping the i-th port in the subgraph
to the i-th port in the replacement.


To convince yourself of that:

### nu_inp: (Node, IncomingPort) -> (HostNode, IncomingPort)

Non-injective case: that’s using a value from the host HUGR multiple
times (copying). This is covered by having a `replacement` HUGR that
copies a value at the input.

Non-surjective case: that’s discarding a value from host. This is
covered by having a `replacement` that discards a value at the input.


### nu_out: (HostNode, IncomingPort) -> (Node, IncomingPort)

Non-injective case: that’s using a value produced by `replacement`
multiple times in the host HUGR. This is replaced by copying the value
in the replacement to multiple outputs

Non-surjective case: that’s discarding a value from `replacement`.
Replaced by not linking that value to the output node.

### "The proof is convincing, but what about the following case?"

Consider a subgraph given by nodes `1` and `2` and consider a wire with
an outgoing port in `1` and two incoming ports, one in `2` and one in
`3`.

```mermaid
flowchart LR
 subgraph G["Subgraph"]
        1
        2
  end
    1 --> 2 & 3
```

The previous version of `nu_out` was able to handle this case, by
specifying that the incoming port in `3` is a boundary output port,
whereas the incoming port in `2` is not. The new version is not, because
the output boundary is now the outgoing port in `1`. Is the proof
broken? No!! In fact, whilst the maps `nu_inp` and `nu_out` can express
this case, the `subgraph` field of `SimpleReplacement` **cannot**
represent this subgraph (it is "non-convex", according to our definition
of convexity). So this PR is in no way a regression in functionality. I
will open a PR to handle this case in `SiblingSubgraph`, at which point
we can update `SimpleReplacement` to handle it too.

BREAKING CHANGE: `SimpleReplacement::new` was replaced by
`SimpleReplacement::try_new`. `SimpleReplacement::{map_host_output,
map_replacement_input}` take and return opposite port directions.