feat: Export macro for hugr serde wrappers with custom extensions (#2209)
The `AsStringEnvelope` wrapper works well enough, but doesn't allow
defining which extension sets to use when loading the Package/Hugrs.

Unfortunately, there is no way to define custom extension registers due
to the way in which `serde_with` works.
The solution here is to add a public macro that implements the required
traits.

This is needed for the ECC sets definition in `tket2`, which encode
hugrs containing tket2-extension operations inside a serde struct.