feat: Symbol applications can leave out prefixes of wildcards. (#2201)
Symbols in `hugr-model` such as functions or constructors can have
parameters whose value can be inferred from context. So far it was
already possible to pass wildcard terms to these parameters to indicate
that they should be inferred. Noticing that redundant parameters often
form a prefix of a symbol's parameter list, this PR adds the ability to
apply a symbol with fewer arguments than it has parameters,
automatically padding the arguments with wildcards in the front.