feat!: Entrypoints in `hugr-py` (#2148)
Adds entrypoint definitions to `hugr-py`. Has similar caveats to #2147.

The main issue here was builders that initialized a hugr for a dataflow
op without knowing its output types. The automatic machinery that wraps
the op in a function definition needs to connect the outputs to the new
function's output once we know the types, so I had to add some extra
machinery to `DataflowOp` to signal back to the hugr that it should
connect things.
This feels a bit hacky, perhaps there's a cleaner way to do it...

Blocked by #2147.

BREAKING CHANGE: Hugrs now define an `entrypoint` in addition to a
module root.