feat: Add boundary edge traversal in SimpleReplacement (#2231)
I believe these methods finally expose the "right" way to think of the
boundary between a replacement and its subgraph, in particular this
won't be broken as we extend `SiblingSubgraph` to express more types of
subgraphs (see issue #2212).

The idea is simple: for any replacement, there must be links between
ports that are just outside the `subgraph` in host, and ports that are
linked to the input and output nodes of the `replacement`. These links
are precisely the edges that will need to be added when the replacement
is applied, between the existing host HUGR and the newly added nodes of
the `replacement`. Specifically these edges link:
- outgoing ports in `host` that are "just before the subgraph" (i.e.
ports linked to an incoming port of `subgraph`) to incoming ports in
`replacement` that are linked to the input node, and
- incoming ports in `host` that are "just after the subgraph" (i.e.
ports linked to an outgoing port of `subgraph`) to outgoing ports in
`replacement` that are linked to the output node.