feat: Add signature map function for DFGs (#2239)
The boundary map was [recently
removed](https://github.com/CQCL/hugr/pull/2208) in `SimpleReplacement`.
Whereas previously, a boundary map could be defined to e.g. permutate
the inputs and outputs between the subgraph and the provided
replacement, the replacement must now match the subgraph function type
as is.

This PR adds a `map_function_type` method on DFG HUGRs that can be used
to reorder and copy/discard inputs and outputs to the DFG pointed by the
entrypoint. For the HUGR to be valid, it is also necessary to update the
signatures and links in any enclosing DFGs in the ancestors of the
entrypoint.

Closes https://github.com/CQCL/hugr/issues/2211

---------

Co-authored-by: Alec Edgington <54802828+cqc-alec@users.noreply.github.com>