fix: Make SumType::Unit(N) equal to SumType::General([(); N]) (#2250)
`SumType` can represent the same sum of empty tuples in two different
ways, but its equality implementation didn't consider them equal.

This PR fixes this by manually implementing `PartialEq`, as well as
`Hash` (to preserve correctness).
We already do that on the python side

https://github.com/CQCL/hugr/blob/940895cf3eae57bb6cae6d13f96bf4f065cb4d2b/hugr-py/src/hugr/tys.py#L311-L312

Ideally we'd make the general-sum-of-empty-tuples unrepresentable, but
since `SumType` is an enum (with public member), we are not able to
avoid arbitrary mutation.