feat!: More helpful error messages in model import (#2264)
This PR adds more helpful error messages to model import to help debug
errors that may occur during the migration to `hugr-model` as the
default serialisation format (see #2260). The errors are still far from
ideal, but should already provide a significant improvement. We hide the
internals of `ImportError` in order to allow for the error messages to
improve without incurring breaking changes. This PR also implements a
remaining `todo!()` concerning importing tuple types.

BREAKING CHANGE: The internals of `ImportError` are hidden.