feat: Add MermaidFormatter to replace RenderConfig (#2275)
To move `PersistentHugr` out of `hugr-core`, I need to be able to
provide custom node labels to the mermaid printer. Whilst adding a field
for this to `RenderConfig` would not be breaking, the problem is that
such a field would have to be an `Option<HashMap>`, which is non-`Copy`
and non-`Hash`. That would be a breaking change to RenderConfig.

I have found a creative way around this here, keeping the Copy-able
`RenderConfig` and adding a `FullRenderConfig` struct with support for
`From` and `TryInto`.

Is this the preferred approach? How would I make a note that this ought
to be simplified ahead of the next breaking release?