feat!: Better error reporting in `hugr-cli`. (#2318)
This PR streamlines the error reporting in `hugr-cli`, displaying error
source trails and setting up the tracing crate. For that, some error
trait implementations were migrated to `thiserror`. The PR also
simplifies the `clap` structures by pulling out the verbosity level from
the individual commands. The commands should use tracing macros instead
of relying on filtering for verbosity themselves.

BREAKING CHANGE: Reorganised `clap` commands and errors in `hugr-cli`
which is only visible when it is used as a library.