fix: update CallGraph and remove_dead_funcs for module-only FuncDefns (#2336)
This should have been part of #2256 but some confusion dates from #2147.

Making a callgraph for only the functions beneath the entrypoint makes
no sense - that's either all the functions (if the entrypoint is the
root) or at most one of them (if the entrypoint is one of the
funcdefns). Of course, we didn't test with any Hugrs that had a
non-module-root entrypoint.... so the sensible thing to do now, seems to
be for callgraph to ignore the entrypoint. (We can optimize it looking
for FuncDefns tho, they are now easy to find.)

Dead function removal allowed specifying starting points for analysis
(top-level FuncDefns) but only if the Hugr's entrypoint was the module.
There seems no good reason not to allow specifying module-level
FuncDefns even if the entrypoint is lower, and similarly, we should use
the entrypoint as one starting point (perhaps among many) for
analysis...

(One could rename `entry_points` used by analysis to "starting_points"
but that would be breaking)