fix!: Fixed bugs in model CFG handling and improved CFG signatures (#2334)
This PR fixes some bugs and a design error related to model CFGs. The
signature of a control flow region and of a dataflow block is now of the
form `(core.ctrl ?inputs ?outputs)` where `?inputs` and `?outputs` are
lists of lists of types. Each item in the outer list corresponds to a
branch and the items in the inner lists of the types of the values that
are transmitted in that branch.

BREAKING CHANGE: The model CFG signature types were changed.