feat(persistent): Add serialisation for CommitStateSpace (#2344)
This PR adds serialisation for the `CommitStateSpace` type, following
the pattern of #2300. To make the serialisation/deserialisation
deterministic, a new `Resolver` is introduced.

Resolvers define a notion of equivalence between commits in the state
space. The simplest resolver is `PointerEqResolver`, which considers
commits to be equivalent if they point to the same data in memory. The
"key" to that commit is the memory pointer, and so is non-deterministic.

The new resolver (`SerdeHashResolver`) determines commit equivalence by
computing a hash of the serialised data. This is not very efficient and
will have to be improved in the future, but it is correct and convenient
to use for the time being. For this purpose, a fast, reproducible and
platform-independent hash is used (`wyhash`).

To accomodate multiple resolves, `CommitStateSpace` and `PersistentHugr`
had to be made generic over the resolver type.

Closes #2299