feat: Deprecate invalidation_set, add invalidated_nodes and SimpleReplacement::invalidation_set (#2358)
Currently a Patch(Verification) has to report its `invalidation_set`
without being able to see the Hugr on which it will act (and without
being able to verify itself against that Hugr - well the set is
meaningless if it doesn't verify/apply, but nonetheless). This would be
really useful in #2290 for one...

So, this PR adds a new method `invalidated_nodes` which is like
`invalidation_set` but takes an `&impl HugrView`; and deprecates the old
`invalidation_set`.

Annoyingly deprecating a method in a trait doesn't flag up
*implementations* of the method, only calls to it. So the breaking
change follow-up to remove invalidation_set (and *require* an
implementation of invalidated_nodes) will be significantly breaking. (If
we didn't have the default now, this would be breaking and the follow-up
still too, however).

For `SimpleReplacement` I've added the method but outside the trait, for
use in PersistentHugr. (It looks a little odd but the deprecation
warning kicks in only if you call the trait method, and the non-trait
method seems to be preferred when both are in scope.)