fix: DeadFuncElimPass+CallGraph w/ non-module-child entrypoint (#2390)
#2336 tried to update CallGraph wrt. entrypoint, but managed to break it
by searching for the entrypoint only among the module-root's children;
if the entrypoint is deeper, this meant it was never scanned or included
in the CallGraph, and dead-func removal would remove....too much.

This fixes that by scanning the entrypoint separately. Note this
produces two nodes in the callgraph: one for whatever function contains
the entrypoint; and one for the entrypoint. The former will have
outgoing edges to all the same targets as the latter, possibly plus some
more. (One could argue the function containing the entrypoint should
"call" the entrypoint, but this is not a proper call, so, no.)

Fixes #2387