fix: Make JSON Schema checks actually work again (#2412)
fixes #2401
* Update generate_schema.py to add an entry to the generated schema
describing what members of `#/$defs` the actual value (i.e. root) should
conform to.
* Add a test that schema validation fails when expected (by
adding/removing fields).
* Some changes to `check_schemas` and related (crate) methods to expose
errors etc.

* Fix various failures in proptests stemming from
* #2309 (the `Arbitrary` instance can generate "TypeArg"s where
"TypeParam"s are expected or vice versa, and these do not confirm to the
JSON schema).
* #2360 - I'm not clear what *well-typed* ArrayOrTermSer::Term can fit
the JSON schema, but our `Arbitrary`s generate plenty that are not
well-typed.
...I've "fixed" the proptests by filtering out non-schema-compliant
instances using `prop_filter`, this is not great, the new situation is
detailed in #2415 for further cleaning up.