feat(per)!: Support empty wires in commits (#2349)
Add support for commits that have empty wires in the replacement graph
(e.g. cancellation of 2x CZ to the identity circuit).

This creates a new difficulty: whereas previously, it was sufficient to
keep track of the nodes in the current PersistentHugr to fully determine
the set of selected commits (each node belongs to a unique commit, so
the set of selected commits is the set of owners of selected nodes), we
must now additionally keep track of the empty commits that were
selected.

This also means that the full construct for a `Commit` must be passed
the set of parent commits (previously, the parent commits were the
commits that the invalidated nodes belonged to).

Closes #2189.
Closes #2241.

BREAKING CHANGE: `SimpleReplacement::{linked_replacement_output,
linked_replacement_input}` take an additional `BoundaryMode` input. Use
`BoundaryMode::default()` for unchanged behaviour.

---------

Co-authored-by: Alan Lawrence <alan.lawrence@cambridgequantum.com>