feat: Add PinnedSubgraph (#2402)
I've factored out the logic that computes the subgraph of a
`PersistentHugr` when constructing a `SimpleReplacement` into a new type
`PinnedSubgraph`. This is required so that commit factories in `tket2`
are able to compute the matched subgraph and extract the matched
subHUGRs.

This is very similar to `SiblingSubgraph`, with the nuance that the same
subgraph can be applied to multiple `PersistentHugr`s, as long as they
all share the same pinned nodes (in particular when the different
`PersistentHugr`s stem from the different expansions of the same
`Walker`). Conversions to and from `SiblingSubgraph`s are provided.